Add files via upload
This machine learning project aims to generate Persian poems by utilizing the works of Hafez, a renowned poet among Persian speakers. The data was processed, and the results demonstrate the meaningfulness of the generated text. However, a significant challenge encountered was the right-to-left format of writing in the Persian language. To overcome this obstacle, the input method was reversed. Despite this complication, the project was successful in generating quality Persian poetry.